fixed narrowing conversion

+++ src/lib/parser.cpp
@@ -59,7 +59,7 @@ bool wrapped_bracks(const std::vector<Token>& tokens, unsigned start, unsigned e
 }
 ASTree::ASNode ASTree::build_array(const std::vector<Token>& tokens, unsigned start, unsigned end){
 	ASNode root{tokens.at(start)};
-	int curr_start{start + 1};
+	int curr_start{static_cast<int>(start + 1)};
 	if(start == end){
 		return root;
 	}
@@ -102,7 +102,7 @@ ASTree::ASNode ASTree::build_array(const std::vector<Token>& tokens, unsigned st
 ASTree::ASNode ASTree::build_call(const std::vector<Token>& tokens, unsigned start, unsigned end){
 	ASNode root{tokens.at(start + 1)};
 	root.add_child(this->buildInfix(tokens, start, start, false));
-	int curr_start{start};
+	int curr_start{static_cast<int>(start)};
 	int bdepth{};
 	int pdepth{};
 	for(unsigned i{start + 2}; i <= end; i++){
